Replace XdsNameResolver with XdsNameResolver2.
