+++ xds/src/main/java/io/grpc/xds/XdsNameResolver.java
@@ -19,44 +19,38 @@ package io.grpc.xds;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.Stopwatch;
-import com.google.common.base.Supplier;
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.Sets;
 import com.google.gson.Gson;
-import com.google.re2j.Pattern;
 import io.grpc.Attributes;
-import io.grpc.EquivalentAddressGroup;
+import io.grpc.CallOptions;
+import io.grpc.InternalConfigSelector;
 import io.grpc.InternalLogId;
+import io.grpc.LoadBalancer.PickSubchannelArgs;
+import io.grpc.Metadata;
 import io.grpc.NameResolver;
 import io.grpc.Status;
 import io.grpc.SynchronizationContext;
-import io.grpc.internal.BackoffPolicy;
 import io.grpc.internal.GrpcUtil;
 import io.grpc.internal.ObjectPool;
 import io.grpc.xds.Bootstrapper.BootstrapInfo;
-import io.grpc.xds.Bootstrapper.ServerInfo;
 import io.grpc.xds.EnvoyProtoData.ClusterWeight;
-import io.grpc.xds.EnvoyProtoData.Node;
 import io.grpc.xds.EnvoyProtoData.Route;
 import io.grpc.xds.EnvoyProtoData.RouteAction;
-import io.grpc.xds.RouteMatch.FractionMatcher;
-import io.grpc.xds.RouteMatch.HeaderMatcher;
-import io.grpc.xds.RouteMatch.PathMatcher;
+import io.grpc.xds.ThreadSafeRandom.ThreadSafeRandomImpl;
 import io.grpc.xds.XdsClient.ConfigUpdate;
 import io.grpc.xds.XdsClient.ConfigWatcher;
-import io.grpc.xds.XdsClient.RefCountedXdsClientObjectPool;
-import io.grpc.xds.XdsClient.XdsChannelFactory;
-import io.grpc.xds.XdsClient.XdsClientFactory;
+import io.grpc.xds.XdsClient.XdsClientPoolFactory;
 import io.grpc.xds.XdsLogger.XdsLogLevel;
-import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.LinkedHashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.ScheduledExecutorService;
-import javax.annotation.Nullable;
+import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * A {@link NameResolver} for resolving gRPC target names with "xds:" scheme.
@@ -68,36 +62,45 @@ import javax.annotation.Nullable;
  */
 final class XdsNameResolver extends NameResolver {
 
+  static final CallOptions.Key<String> CLUSTER_SELECTION_KEY =
+      CallOptions.Key.create("io.grpc.xds.CLUSTER_SELECTION_KEY");
+
   private final XdsLogger logger;
   private final String authority;
-  private final XdsChannelFactory channelFactory;
-  private final SynchronizationContext syncContext;
-  private final ScheduledExecutorService timeService;
   private final ServiceConfigParser serviceConfigParser;
-  private final BackoffPolicy.Provider backoffPolicyProvider;
-  private final Supplier<Stopwatch> stopwatchSupplier;
+  private final SynchronizationContext syncContext;
   private final Bootstrapper bootstrapper;
+  private final XdsClientPoolFactory xdsClientPoolFactory;
+  private final ThreadSafeRandom random;
+  private final ConcurrentMap<String, AtomicInteger> clusterRefs = new ConcurrentHashMap<>();
+  private final ConfigSelector configSelector = new ConfigSelector();
 
-  @Nullable
+  private volatile List<Route> routes = Collections.emptyList();
+  private Listener2 listener;
   private ObjectPool<XdsClient> xdsClientPool;
-  @Nullable
   private XdsClient xdsClient;
+
+  XdsNameResolver(String name,
+      ServiceConfigParser serviceConfigParser,
+      SynchronizationContext syncContext,
+      XdsClientPoolFactory xdsClientPoolFactory) {
+    this(name, serviceConfigParser, syncContext, Bootstrapper.getInstance(),
+        xdsClientPoolFactory, ThreadSafeRandomImpl.instance);
+  }
 
   XdsNameResolver(
       String name,
-      Args args,
-      BackoffPolicy.Provider backoffPolicyProvider,
-      Supplier<Stopwatch> stopwatchSupplier,
-      XdsChannelFactory channelFactory,
-      Bootstrapper bootstrapper) {
+      ServiceConfigParser serviceConfigParser,
+      SynchronizationContext syncContext,
+      Bootstrapper bootstrapper,
+      XdsClientPoolFactory xdsClientPoolFactory,
+      ThreadSafeRandom random) {
     authority = GrpcUtil.checkAuthority(checkNotNull(name, "name"));
-    this.channelFactory = checkNotNull(channelFactory, "channelFactory");
-    this.syncContext = checkNotNull(args.getSynchronizationContext(), "syncContext");
-    this.timeService = checkNotNull(args.getScheduledExecutorService(), "timeService");
-    this.serviceConfigParser = checkNotNull(args.getServiceConfigParser(), "serviceConfigParser");
-    this.backoffPolicyProvider = checkNotNull(backoffPolicyProvider, "backoffPolicyProvider");
-    this.stopwatchSupplier = checkNotNull(stopwatchSupplier, "stopwatchSupplier");
+    this.serviceConfigParser = checkNotNull(serviceConfigParser, "serviceConfigParser");
+    this.syncContext = checkNotNull(syncContext, "syncContext");
     this.bootstrapper = checkNotNull(bootstrapper, "bootstrapper");
+    this.xdsClientPoolFactory = checkNotNull(xdsClientPoolFactory, "xdsClientPoolFactory");
+    this.random = checkNotNull(random, "random");
     logger = XdsLogger.withLogId(InternalLogId.allocate("xds-resolver", name));
     logger.log(XdsLogLevel.INFO, "Created resolver for {0}", name);
   }
@@ -109,78 +112,242 @@ final class XdsNameResolver extends NameResolver {
 
   @Override
   public void start(Listener2 listener) {
+    this.listener = checkNotNull(listener, "listener");
     BootstrapInfo bootstrapInfo;
     try {
       bootstrapInfo = bootstrapper.readBootstrap();
     } catch (Exception e) {
-      listener.onError(Status.UNAVAILABLE.withDescription("Failed to bootstrap").withCause(e));
-      return;
-    }
-    final List<ServerInfo> serverList = bootstrapInfo.getServers();
-    final Node node = bootstrapInfo.getNode();
-    if (serverList.isEmpty()) {
       listener.onError(
-          Status.UNAVAILABLE.withDescription("No management server provided by bootstrap"));
+          Status.UNAVAILABLE.withDescription("Failed to load xDS bootstrap").withCause(e));
       return;
     }
-
-    XdsClientFactory xdsClientFactory = new XdsClientFactory() {
-      @Override
-      XdsClient createXdsClient() {
-        return
-            new XdsClientImpl(
-                authority,
-                serverList,
-                channelFactory,
-                node,
-                syncContext,
-                timeService,
-                backoffPolicyProvider,
-                stopwatchSupplier);
-      }
-    };
-    xdsClientPool = new RefCountedXdsClientObjectPool(xdsClientFactory);
+    xdsClientPool = xdsClientPoolFactory.newXdsClientObjectPool(bootstrapInfo);
     xdsClient = xdsClientPool.getObject();
-    xdsClient.watchConfigData(authority, new ConfigWatcherImpl(listener));
+    xdsClient.watchConfigData(authority, new ConfigWatcherImpl());
   }
 
-  private class ConfigWatcherImpl implements ConfigWatcher {
-
-    final Listener2 listener;
-
-    ConfigWatcherImpl(Listener2 listener) {
-      this.listener = listener;
+  @Override
+  public void shutdown() {
+    logger.log(XdsLogLevel.INFO, "Shutdown");
+    if (xdsClient != null) {
+      xdsClient = xdsClientPool.returnObject(xdsClient);
     }
+  }
+
+  @VisibleForTesting
+  static Map<String, ?> generateServiceConfigWithMethodTimeoutConfig(long timeoutNano) {
+    String timeout = timeoutNano / 1_000_000_000.0 + "s";
+    Map<String, Object> methodConfig = new HashMap<>();
+    methodConfig.put(
+        "name", Collections.singletonList(Collections.emptyMap()));
+    methodConfig.put("timeout", timeout);
+    return Collections.singletonMap(
+        "methodConfig", Collections.singletonList(Collections.unmodifiableMap(methodConfig)));
+  }
+
+  @VisibleForTesting
+  static Map<String, ?> generateServiceConfigWithLoadBalancingConfig(Collection<String> clusters) {
+    Map<String, Object> childPolicy = new HashMap<>();
+    for (String cluster : clusters) {
+      List<Map<String, Map<String, String>>> lbPolicy =
+          Collections.singletonList(
+              Collections.singletonMap(
+                  "cds_experimental", Collections.singletonMap("cluster", cluster)));
+      childPolicy.put(cluster, Collections.singletonMap("lbPolicy", lbPolicy));
+    }
+    return Collections.singletonMap("loadBalancingConfig",
+        Collections.singletonList(
+            Collections.singletonMap(
+                "cluster_manager_experimental", Collections.singletonMap(
+                    "childPolicy", Collections.unmodifiableMap(childPolicy)))));
+  }
+
+  @VisibleForTesting
+  XdsClient getXdsClient() {
+    return xdsClient;
+  }
+
+  private void updateResolutionResult() {
+    Map<String, ?> rawServiceConfig =
+        generateServiceConfigWithLoadBalancingConfig(clusterRefs.keySet());
+    if (logger.isLoggable(XdsLogLevel.INFO)) {
+      logger.log(
+          XdsLogLevel.INFO, "Generated service config:\n{0}", new Gson().toJson(rawServiceConfig));
+    }
+    ConfigOrError parsedServiceConfig = serviceConfigParser.parseServiceConfig(rawServiceConfig);
+    Attributes attrs =
+        Attributes.newBuilder()
+            .set(XdsAttributes.XDS_CLIENT_POOL, xdsClientPool)
+            .set(InternalConfigSelector.KEY, configSelector)
+            .build();
+    ResolutionResult result =
+        ResolutionResult.newBuilder()
+            .setAttributes(attrs)
+            .setServiceConfig(parsedServiceConfig)
+            .build();
+    listener.onResult(result);
+  }
+
+  private final class ConfigSelector extends InternalConfigSelector {
+    @Override
+    public Result selectConfig(PickSubchannelArgs args) {
+      // Index ASCII headers by keys.
+      Map<String, Iterable<String>> asciiHeaders = new HashMap<>();
+      Metadata headers = args.getHeaders();
+      for (String headerName : headers.keys()) {
+        if (headerName.endsWith(Metadata.BINARY_HEADER_SUFFIX)) {
+          continue;
+        }
+        Metadata.Key<String> key = Metadata.Key.of(headerName, Metadata.ASCII_STRING_MARSHALLER);
+        asciiHeaders.put(headerName, headers.getAll(key));
+      }
+      String cluster = null;
+      Route selectedRoute = null;
+      do {
+        for (Route route : routes) {
+          if (route.getRouteMatch().matches(
+              "/" + args.getMethodDescriptor().getFullMethodName(), asciiHeaders)) {
+            selectedRoute = route;
+            break;
+          }
+        }
+        if (selectedRoute == null) {
+          return Result.forError(
+              Status.UNAVAILABLE.withDescription("Could not find xDS route matching RPC"));
+        }
+        RouteAction action = selectedRoute.getRouteAction();
+        if (action.getCluster() != null) {
+          cluster = action.getCluster();
+        } else if (action.getWeightedCluster() != null) {
+          int totalWeight = 0;
+          for (ClusterWeight weightedCluster : action.getWeightedCluster()) {
+            totalWeight += weightedCluster.getWeight();
+          }
+          int select = random.nextInt(totalWeight);
+          int accumulator = 0;
+          for (ClusterWeight weightedCluster : action.getWeightedCluster()) {
+            accumulator += weightedCluster.getWeight();
+            if (select < accumulator) {
+              cluster = weightedCluster.getName();
+              break;
+            }
+          }
+        }
+      } while (!retainCluster(cluster));
+      // TODO(chengyuanzhang): avoid service config generation and parsing for each call.
+      Map<String, ?> rawServiceConfig =
+          generateServiceConfigWithMethodTimeoutConfig(
+              selectedRoute.getRouteAction().getTimeoutNano());
+      if (logger.isLoggable(XdsLogLevel.INFO)) {
+        logger.log(XdsLogLevel.INFO,
+            "Generated service config (method config):\n{0}", new Gson().toJson(rawServiceConfig));
+      }
+      ConfigOrError parsedServiceConfig = serviceConfigParser.parseServiceConfig(rawServiceConfig);
+      Object config = parsedServiceConfig.getConfig();
+      if (config == null) {
+        releaseCluster(cluster);
+        return Result.forError(
+            parsedServiceConfig.getError().augmentDescription(
+                "Failed to parse service config (method config)"));
+      }
+      final String finalCluster = cluster;
+      class SelectionCompleted implements Runnable {
+        @Override
+        public void run() {
+          releaseCluster(finalCluster);
+        }
+      }
+
+      return
+          Result.newBuilder()
+              .setCallOptions(args.getCallOptions().withOption(CLUSTER_SELECTION_KEY, cluster))
+              .setConfig(config)
+              .setCommittedCallback(new SelectionCompleted())
+              .build();
+    }
+
+    private boolean retainCluster(String cluster) {
+      AtomicInteger refCount = clusterRefs.get(cluster);
+      if (refCount == null) {
+        return false;
+      }
+      int count;
+      do {
+        count = refCount.get();
+        if (count == 0) {
+          return false;
+        }
+      } while (!refCount.compareAndSet(count, count + 1));
+      return true;
+    }
+
+    private void releaseCluster(final String cluster) {
+      int count = clusterRefs.get(cluster).decrementAndGet();
+      if (count == 0) {
+        syncContext.execute(new Runnable() {
+          @Override
+          public void run() {
+            if (clusterRefs.get(cluster).get() == 0) {
+              clusterRefs.remove(cluster);
+              updateResolutionResult();
+            }
+          }
+        });
+      }
+    }
+  }
+
+  // https://github.com/google/error-prone/issues/1767
+  @SuppressWarnings("ModifyCollectionInEnhancedForLoop")
+  private class ConfigWatcherImpl implements ConfigWatcher {
+    private Set<String> existingClusters;
 
     @Override
     public void onConfigChanged(ConfigUpdate update) {
-      logger.log(
-          XdsLogLevel.INFO,
-          "Received config update with {0} routes from xDS client {1}",
-          update.getRoutes().size(),
-          xdsClient);
-      Map<String, ?> rawLbConfig = generateXdsRoutingRawConfig(update.getRoutes());
-      Map<String, ?> serviceConfig =
-          ImmutableMap.of("loadBalancingConfig", ImmutableList.of(rawLbConfig));
-      if (logger.isLoggable(XdsLogLevel.INFO)) {
-        logger.log(
-            XdsLogLevel.INFO,
-            "Generated service config:\n{0}",
-            new Gson().toJson(serviceConfig));
+      Set<String> clusters = new HashSet<>();
+      for (Route route : update.getRoutes()) {
+        RouteAction action = route.getRouteAction();
+        if (action.getCluster() != null) {
+          clusters.add(action.getCluster());
+        } else if (action.getWeightedCluster() != null) {
+          for (ClusterWeight weighedCluster : action.getWeightedCluster()) {
+            clusters.add(weighedCluster.getName());
+          }
+        }
       }
-
-      Attributes attrs =
-          Attributes.newBuilder()
-              .set(XdsAttributes.XDS_CLIENT_POOL, xdsClientPool)
-              .build();
-      ConfigOrError parsedServiceConfig = serviceConfigParser.parseServiceConfig(serviceConfig);
-      ResolutionResult result =
-          ResolutionResult.newBuilder()
-              .setAddresses(ImmutableList.<EquivalentAddressGroup>of())
-              .setAttributes(attrs)
-              .setServiceConfig(parsedServiceConfig)
-              .build();
-      listener.onResult(result);
+      Set<String> addedClusters =
+          existingClusters == null ? clusters : Sets.difference(clusters, existingClusters);
+      Set<String> deletedClusters =
+          existingClusters == null
+              ? Collections.<String>emptySet() : Sets.difference(existingClusters, clusters);
+      existingClusters = clusters;
+      boolean shouldUpdateResult = false;
+      for (String cluster : addedClusters) {
+        if (clusterRefs.containsKey(cluster)) {
+          clusterRefs.get(cluster).incrementAndGet();
+        } else {
+          clusterRefs.put(cluster, new AtomicInteger(1));
+          shouldUpdateResult = true;
+        }
+      }
+      // Update service config to include newly added clusters.
+      if (shouldUpdateResult) {
+        updateResolutionResult();
+      }
+      // Make newly added clusters selectable by config selector and deleted clusters no longer
+      // selectable.
+      routes = update.getRoutes();
+      shouldUpdateResult = false;
+      for (String cluster : deletedClusters) {
+        int count = clusterRefs.get(cluster).decrementAndGet();
+        if (count == 0) {
+          clusterRefs.remove(cluster);
+          shouldUpdateResult = true;
+        }
+      }
+      if (shouldUpdateResult) {
+        updateResolutionResult();
+      }
     }
 
     @Override
@@ -191,6 +358,7 @@ final class XdsNameResolver extends NameResolver {
       ResolutionResult result =
           ResolutionResult.newBuilder()
               .setServiceConfig(parsedServiceConfig)
+              // let channel take action for no config selector
               .build();
       listener.onResult(result);
     }
@@ -201,154 +369,6 @@ final class XdsNameResolver extends NameResolver {
           XdsLogLevel.WARNING,
           "Received error from xDS client {0}: {1}", xdsClient, error.getDescription());
       listener.onError(error);
-    }
-  }
-
-  @VisibleForTesting
-  static ImmutableMap<String, ?> generateXdsRoutingRawConfig(List<Route> routes) {
-    List<Object> rawRoutes = new ArrayList<>();
-    Map<String, Object> rawActions = new LinkedHashMap<>();
-    Map<RouteAction, String> existingActions = new HashMap<>();
-    for (Route route : routes) {
-      RouteAction routeAction = route.getRouteAction();
-      String actionName;
-      Map<String, ?> actionPolicy;
-      if (existingActions.containsKey(routeAction)) {
-        actionName = existingActions.get(routeAction);
-      } else {
-        if (routeAction.getCluster() != null) {
-          actionName = "cds:" + routeAction.getCluster();
-          actionPolicy = generateCdsRawConfig(routeAction.getCluster());
-        } else {
-          StringBuilder sb = new StringBuilder("weighted:");
-          List<ClusterWeight> clusterWeights = routeAction.getWeightedCluster();
-          for (ClusterWeight clusterWeight : clusterWeights) {
-            sb.append(clusterWeight.getName()).append('_');
-          }
-          sb.append(routeAction.hashCode());
-          actionName = sb.toString();
-          if (rawActions.containsKey(actionName)) {
-            // Just in case of hash collision, append existingActions.size() to make actionName
-            // unique. However, in case of collision, when new ConfigUpdate is received, actions
-            // and actionNames might be associated differently from the previous update, but it
-            // is just suboptimal and won't cause a problem.
-            actionName = actionName + "_" + existingActions.size();
-          }
-          actionPolicy = generateWeightedTargetRawConfig(clusterWeights);
-        }
-        existingActions.put(routeAction, actionName);
-        List<?> childPolicies = ImmutableList.of(actionPolicy);
-        rawActions.put(actionName, ImmutableMap.of("childPolicy", childPolicies));
-      }
-      ImmutableMap<String, ?> configRoute = convertToRawRoute(route.getRouteMatch(), actionName);
-      rawRoutes.add(configRoute);
-    }
-    return ImmutableMap.of(
-        XdsLbPolicies.XDS_ROUTING_POLICY_NAME,
-        ImmutableMap.of(
-            "route", Collections.unmodifiableList(rawRoutes),
-            "action", Collections.unmodifiableMap(rawActions)));
-  }
-
-  @VisibleForTesting
-  static ImmutableMap<String, ?> convertToRawRoute(RouteMatch routeMatch, String actionName) {
-    ImmutableMap.Builder<String, Object> configRouteBuilder = new ImmutableMap.Builder<>();
-
-    PathMatcher pathMatcher = routeMatch.getPathMatch();
-    String path = pathMatcher.getPath();
-    String prefix = pathMatcher.getPrefix();
-    Pattern regex = pathMatcher.getRegEx();
-    if (path != null) {
-      configRouteBuilder.put("path", path);
-    }
-    if (prefix != null) {
-      configRouteBuilder.put("prefix", prefix);
-    }
-    if (regex != null) {
-      configRouteBuilder.put("regex", regex.pattern());
-    }
-
-    ImmutableList.Builder<Object> rawHeaderMatcherListBuilder = new ImmutableList.Builder<>();
-    List<HeaderMatcher> headerMatchers = routeMatch.getHeaderMatchers();
-    for (HeaderMatcher headerMatcher : headerMatchers) {
-      ImmutableMap.Builder<String, Object> rawHeaderMatcherBuilder = new ImmutableMap.Builder<>();
-      rawHeaderMatcherBuilder.put("name", headerMatcher.getName());
-      String exactMatch = headerMatcher.getExactMatch();
-      Pattern regexMatch = headerMatcher.getRegExMatch();
-      HeaderMatcher.Range rangeMatch = headerMatcher.getRangeMatch();
-      Boolean presentMatch = headerMatcher.getPresentMatch();
-      String prefixMatch = headerMatcher.getPrefixMatch();
-      String suffixMatch = headerMatcher.getSuffixMatch();
-      if (exactMatch != null) {
-        rawHeaderMatcherBuilder.put("exactMatch", exactMatch);
-      }
-      if (regexMatch != null) {
-        rawHeaderMatcherBuilder.put("regexMatch", regexMatch.pattern());
-      }
-      if (rangeMatch != null) {
-        rawHeaderMatcherBuilder
-            .put(
-                "rangeMatch",
-                ImmutableMap.of("start", rangeMatch.getStart(), "end", rangeMatch.getEnd()));
-      }
-      if (presentMatch != null) {
-        rawHeaderMatcherBuilder.put("presentMatch", presentMatch);
-      }
-      if (prefixMatch != null) {
-        rawHeaderMatcherBuilder.put("prefixMatch", prefixMatch);
-      }
-      if (suffixMatch != null) {
-        rawHeaderMatcherBuilder.put("suffixMatch", suffixMatch);
-      }
-      rawHeaderMatcherBuilder.put("invertMatch", headerMatcher.isInvertedMatch());
-      rawHeaderMatcherListBuilder.add(rawHeaderMatcherBuilder.build());
-    }
-    ImmutableList<?> rawHeaderMatchers = rawHeaderMatcherListBuilder.build();
-    if (!rawHeaderMatchers.isEmpty()) {
-      configRouteBuilder.put("headers", rawHeaderMatchers);
-    }
-
-    FractionMatcher matchFraction = routeMatch.getFractionMatch();
-    if (matchFraction != null) {
-      configRouteBuilder
-          .put(
-              "matchFraction",
-              ImmutableMap.of(
-                  "numerator", matchFraction.getNumerator(),
-                  "denominator", matchFraction.getDenominator()));
-    }
-
-    configRouteBuilder.put("action", actionName);
-    return configRouteBuilder.build();
-  }
-
-  @VisibleForTesting
-  static ImmutableMap<String, ?> generateWeightedTargetRawConfig(
-      List<ClusterWeight> clusterWeights) {
-    Map<String, Object> targets = new LinkedHashMap<>();
-    for (ClusterWeight clusterWeight : clusterWeights) {
-      Map<String, ?> childPolicy = generateCdsRawConfig(clusterWeight.getName());
-      Map<String, ?> weightedConfig = ImmutableMap.of(
-          "weight",
-          (double) clusterWeight.getWeight(),
-          "childPolicy",
-          ImmutableList.of(childPolicy));
-      targets.put(clusterWeight.getName(), weightedConfig);
-    }
-    return ImmutableMap.of(
-        XdsLbPolicies.WEIGHTED_TARGET_POLICY_NAME,
-        ImmutableMap.of("targets", targets));
-  }
-
-  private static ImmutableMap<String, ?> generateCdsRawConfig(String clusterName) {
-    return ImmutableMap.of(XdsLbPolicies.CDS_POLICY_NAME, ImmutableMap.of("cluster", clusterName));
-  }
-
-  @Override
-  public void shutdown() {
-    logger.log(XdsLogLevel.INFO, "Shutdown");
-    if (xdsClient != null) {
-      xdsClient = xdsClientPool.returnObject(xdsClient);
     }
   }
 }
